Reorganize bluetooth integration to prepare for remote and multi-adapter support
